create nodestable
